Change scale and shift parameters to double

+++ kleidicv/include/kleidicv/arithmetics/scale.h
@@ -11,11 +11,12 @@
 
 namespace kleidicv::neon {
 
-std::array<uint8_t, 256> precalculate_scale_table_u8(float scale, float shift);
+std::array<uint8_t, 256> precalculate_scale_table_u8(double scale,
+                                                     double shift);
 
 kleidicv_error_t scale_with_precalculated_table(
     const uint8_t *src, size_t src_stride, uint8_t *dst, size_t dst_stride,
-    size_t width, size_t height, float scale, float shift,
+    size_t width, size_t height, double scale, double shift,
     const std::array<uint8_t, 256> &precalculated_table);
 
 }  // namespace kleidicv::neon
